Sketch basic core fragment
Sketch an initial core language that removes incidental complexity from the Rust AST
the goal is to desugar to a small core language that will be amenable to a small type
checker useful for experimentation.

The end goal will be to desugar, collect definitions, then type check in as few passes
as possible. I believe by taking inspiration from bi-directional type checking it will
be possible to dramatically simplify/collapse the compiler passes.